Added in-app directory feature, added to and edited MasterMenuItem list, added to and edited detail views, added Content folder which contains new content pages.

+++ TestProj/Models/Constants.cs
@@ -10,7 +10,7 @@ namespace TestProj.Models
         public static Color DarkModeBackgroundColor = Color.FromArgb(30, 30, 30);
         public static Color MainTextColor = Color.White;
 
-        public static int LoginIconHeight = 120;
+        public static double LoginIconHeight = 120;
 
 
         //------Login------------
